requirements.txt
django
pillow

# DataBase
psycopg2-binary

# restframework
djangorestframework
markdown
django-filter

# documentation
drf-yasg[validation]